+++ tests-trybuild/tests/riscv-rt/core_interrupt/fail_empty_macro.rs
@@ -1,4 +1,0 @@
-#[riscv_rt::core_interrupt]
-fn my_interrupt() {}
-
-fn main() {}