+++ tests-trybuild/tests/riscv-rt/exception/fail_empty_args.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::exception]
+pub const async extern "Rust" fn my_exception<T>(_a: T, _b: usize, _c: ...) -> usize {}
+
+fn main() {}