+++ tests-trybuild/tests/riscv-rt/exception/pass_exception.rs
@@ -4,10 +4,10 @@ use riscv::interrupt::Exception::*;
 fn simple_exception() {}
 
 #[riscv_rt::exception(LoadFault)]
-fn unmutable_exception(_trap_frame: &riscv_rt::TrapFrame) {}
+fn immutable_exception(_trap_frame: &riscv_rt::TrapFrame) {}
 
 #[riscv_rt::exception(StoreMisaligned)]
-fn mutable_exception(_trap_frame: &mut riscv_rt::TrapFrame) {}
+unsafe fn mutable_exception(_trap_frame: &mut riscv_rt::TrapFrame) {}
 
 #[riscv_rt::exception(StoreFault)]
 fn no_return_exception(_trap_frame: &mut riscv_rt::TrapFrame) -> ! {