+++ tests-trybuild/tests/riscv-rt/external_interrupt/fail_empty_args.rs
@@ -1,0 +1,4 @@
+#[riscv_rt::external_interrupt]
+pub const async extern "Rust" fn my_interrupt<T>(_a: T, _b: ...) -> usize {}
+
+fn main() {}